+++ changelog.md
@@ -2,3 +2,6 @@
 
 ## Changes
 - Updated Sky Aesthetics (1.4.4 => 1.5.0)
+
+## Fix
+- Fix fabric mob effect